Move provider resolution to internal/provider (#361)
Migrate resolveProvider/resolveNamedProvider/resolveBaseURL into
internal/provider as exported ResolveProvider/ResolveNamedProvider/
ResolveBaseURL, injecting os.Getenv as an env func(string) string
(last param, matching ResolveSpecialProvider). Move printProviderHelp
into internal/cli as PrintProviderHelp. Relocate the resolution and
base-url tests to internal/provider; the help test to internal/cli.
cmd/pigo now only calls the new packages. Resolution precedence and
--help output are unchanged.
